🐛🤖 Move env override to a pre-import shim so it propagates
The in-bundle approach (top-level await in src/lib/server/env-override.ts
imported from hooks.server.ts) was too late: SvelteKit's manifest
evaluates database.ts before hooks.server.ts, so MONGODB_URL is captured
from process.env before the override mutation. Confirmed in production
where the page kept showing the bootstrap host even though the log said
"applied 2 env var override(s)".

Moving the discovery logic to scripts/env-override-preimport.mjs and
invoking it via `node --import` guarantees it runs before any SvelteKit
module evaluates. The TS file becomes a thin constants-only module used
by the admin UI; hooks.server.ts no longer imports it.

Launch commands updated:
- package.json run-production
- entrypoint.sh (Docker)
- README bare + pm2 examples
- scripts/build-release.sh copies the shim into the artifact

Requires Node 20+ (or 18.20+) for the --import flag. Dropping the flag
silently disables the ALLOW_ENV_OVERRIDE feature; be-BOP runs normally.
