+++ README.md
@@ -70,21 +70,23 @@ You can also set the following environment variables to allow SSO. Set your redi
 
 ```shell
 pnpm run build
-node --enable-source-maps build/index.js
+node --enable-source-maps --import ./scripts/env-override-preimport.mjs build/index.js
 
 # If behind a reverse proxy, you can use the following config:
-# ADDRESS_HEADER=X-Forwarded-For XFF_DEPTH=1 node build/index.js
+# ADDRESS_HEADER=X-Forwarded-For XFF_DEPTH=1 node --import ./scripts/env-override-preimport.mjs build/index.js
 ```
+
+The `--import ./scripts/env-override-preimport.mjs` flag enables the `ALLOW_ENV_OVERRIDE` feature (admin-uploaded MongoDB switch from `/admin/be-bop`). It is a no-op when the flag is unset. Requires Node 20+ (or 18.20+). You can drop the flag entirely if you do not need this feature.
 
 You can set the `PORT` environment variable to change from the default 3000 port to another port.
 
 You can also use [pm2](https://pm2.keymetrics.io/docs/usage/quick-start/) to manage your node application, and run it on multiple cores.
 
 ```shell
-NODE_OPTIONS=--enable-source-maps pm2 start --name bebop --update-env build/index.js
+NODE_OPTIONS="--enable-source-maps --import ./scripts/env-override-preimport.mjs" pm2 start --name bebop --update-env build/index.js
 
 # If behind a reverse proxy, you can use the following config:
-# NODE_OPTIONS=--enable-source-maps ADDRESS_HEADER=X-Forwarded-For XFF_DEPTH=1 pm2 start --name bebop --update-env build/index.js
+# NODE_OPTIONS="--enable-source-maps --import ./scripts/env-override-preimport.mjs" ADDRESS_HEADER=X-Forwarded-For XFF_DEPTH=1 pm2 start --name bebop --update-env build/index.js
 ```
 
 Note: for uploading large payloads you may want to set the `BODY_SIZE_LIMIT=20000000` environment variable to allow 20MB payloads for example. It should not be needed for normal usage.
