+++ entrypoint.sh
@@ -60,4 +60,4 @@ if ! grep -q "^${S3_KEY_SECRET}$" ${ENV_LOCAL_PATH}; then
 fi
 
 ORIGIN="${ORIGIN:-http://localhost:3000}" PUBLIC_VERSION=$(cat ./.git-commit-hash) pnpm run build
-PORT=3000 node --enable-source-maps build/index.js
+PORT=3000 node --enable-source-maps --import ./scripts/env-override-preimport.mjs build/index.js
