+++ scripts/build-release.sh
@@ -44,6 +44,8 @@ cp package.json "$OUTPUT_DIR/$ARTIFACT_NAME/package.json"
 cp pnpm-lock.yaml "$OUTPUT_DIR/$ARTIFACT_NAME/pnpm-lock.yaml"
 cp -r ./assets "$OUTPUT_DIR/$ARTIFACT_NAME/assets"
 cp -r ./docs "$OUTPUT_DIR/$ARTIFACT_NAME/docs"
+mkdir -p "$OUTPUT_DIR/$ARTIFACT_NAME/scripts"
+cp ./scripts/env-override-preimport.mjs "$OUTPUT_DIR/$ARTIFACT_NAME/scripts/env-override-preimport.mjs"
 
 echo "✅ Release artifact created successfully!"
 echo "📁 Directory: $OUTPUT_DIR/$ARTIFACT_NAME"
